requirements.txt
streamlit
Pillow
numpy
imagehash
opencv-python-headless
scikit-image
torch
torchvision
tqdm
scikit-learn
joblib
